+++ marlowe-actus/actus-core/src/Actus/Domain/Schedule.hs
@@ -3,7 +3,6 @@
 {-# LANGUAGE DerivingStrategies #-}
 module Actus.Domain.Schedule where
 
-import Actus.Domain.BusinessEvents (EventType)
 import Data.Aeson.Types (FromJSON, ToJSON)
 import Data.Time (LocalTime)
 import GHC.Generics (Generic)
@@ -19,18 +18,3 @@ mkShiftedDay :: LocalTime -> ShiftedDay
 mkShiftedDay d = ShiftedDay d d
 
 type ShiftedSchedule = [ShiftedDay]
-
-data CashFlow a = CashFlow
-  { tick               :: Integer,
-    cashContractId     :: String,
-    cashParty          :: String,
-    cashCounterParty   :: String,
-    cashPaymentDay     :: LocalTime,
-    cashCalculationDay :: LocalTime,
-    cashEvent          :: EventType,
-    amount             :: a,
-    notional           :: a,
-    currency           :: String
-  }
-  deriving stock (Show, Eq, Generic)
-  deriving anyclass (FromJSON, ToJSON)
